New interfaces added/refactored.
