+++ Source/SimpleForex.Core/Contracts/Commands/ICommandBase.cs
@@ -1,0 +1,7 @@
+namespace SimpleForex.Core.Contracts.Commands
+{
+    /// <summary>
+    /// Represents a base command interface.
+    /// </summary>
+    public interface ICommandBase { }
+}
