+++ Source/SimpleForex.Core/Contracts/Factories/ICommandFactory.cs
@@ -1,4 +1,6 @@
-﻿namespace SimpleForex.Core.Contracts
+using SimpleForex.Core.Contracts.Commands;
+
+namespace SimpleForex.Core.Contracts.Factories
 {
     public interface ICommandFactory
     {
