+++ Source/SimpleForex.Core/Contracts/Factories/IQueryFactory.cs
@@ -1,12 +1,12 @@
-﻿namespace SimpleForex.Core.Contracts
+namespace SimpleForex.Core.Contracts.Factories
 {
     /// <summary>
-    /// Create an instance of an async query.
+    /// Create an instance of a query.
     /// </summary>
     public interface IQueryFactory
     {
         /// <summary>
-        /// Get an instance of an async query.
+        /// Get an instance of a query.
         /// </summary>
         /// <typeparam name="TQuery">Query type.</typeparam>
         TQuery MakeQuery<TQuery>() where TQuery : IQueryBase;
