--- Source/SimpleForex.Core/Contracts/ICurrencyConsultService.cs
@@ -1,7 +1,0 @@
-namespace SimpleForex.Core.Contracts
-{
-    public interface ICurrencyConsultService
-    {
-        decimal Consult();
-    }
-}
