+++ Source/SimpleForex.Core/Contracts/IEntity.cs
@@ -1,5 +1,3 @@
-﻿using System;
-
 namespace SimpleForex.Core.Contracts
 {
     /// <summary>
@@ -12,30 +10,5 @@ namespace SimpleForex.Core.Contracts
         /// The entity unique identifier.
         /// </summary>
         T Id { get; set; }
-
-        /// <summary>
-        /// The unique identifier for the entity creator.
-        /// </summary>
-        T CreatedBy { get; set; }
-
-        /// <summary>
-        /// The date and time where the entity was created.
-        /// </summary>
-        DateTime CreatedDate { get; set; }
-
-        /// <summary>
-        /// The unique identifier for the user who update the entity.
-        /// </summary>
-        T UpdatedBy { get; set; }
-
-        /// <summary>
-        /// The date and time where the entity was updated.
-        /// </summary>
-        DateTime UpdatedDate { get; set; }
-
-        /// <summary>
-        /// Indicates whether or not the entity is deleted.
-        /// </summary>
-        string DeleteFlag { get; set; }
     }
 }
