+++ Source/SimpleForex.Core/Contracts/Queries/IQueryBase.cs
@@ -1,0 +1,7 @@
+namespace SimpleForex.Core.Contracts
+{
+    /// <summary>
+    /// Represent a general query interface.
+    /// </summary>
+    public interface IQueryBase { }
+}
