+++ Source/SimpleForex.Core/Contracts/Services/IService.cs
@@ -1,0 +1,16 @@
+namespace SimpleForex.Core.Contracts.Services
+{
+
+    /// <summary>
+    /// Represents a base service interface.
+    /// </summary>
+    /// <typeparam name="TResult">Service result.</typeparam>
+    public interface IService<TParam, TResult> : IServiceBase where TResult : IEntity<int>
+    {
+        /// <summary>
+        /// Runs the service returning a specified result when it finishes.
+        /// </summary>
+        /// <returns><typeparamref name="TResult"/></returns>
+        TResult RunService(TParam param);
+    }
+}
