+++ Source/SimpleForex.Core/Contracts/Services/IServiceBase.cs
@@ -1,0 +1,7 @@
+namespace SimpleForex.Core.Contracts.Services
+{
+    /// <summary>
+    /// Represents a general service interface.
+    /// </summary>
+    public interface IServiceBase { }
+}
